+++ test/integration/default/controls/default_control.rb
@@ -16,9 +16,7 @@ control 'default' do
     end
 
     describe yum.repo "#{name}-debuginfo" do
-      it { should exist }
-      it { should_not be_enabled }
-      its('mirrors') { should cmp "https://mirrors.almalinux.org/mirrorlist/#{os_release}/#{name}-debuginfo/" }
+      it { should_not exist }
     end
   end